requirements.txt
python-telegram-bot==20.7
flask==3.0.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
paramiko==3.4.0
uuid==1.30
requests==2.31.0
persiantools==4.1.1
pytz==2024.1
